more clean up

+++ src/system_call_names.rs
@@ -3,7 +3,7 @@ use anyhow::Context;
 
 pub fn get_syscall_name(number: usize) -> anyhow::Result<&'static str> {
     let name = SYSTEM_CALL_NAMES
-        .get(number as usize)
+        .get(number)
         .with_context(|| context!("System call number {} out of bounds.", number))?;
     Ok(*name)
 }
